requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
python-dateutil>=2.8.2
